Move custom logger class in l.psql
